+++ src/WebJobs.Script.WebHost/Management/IMeshServiceClient.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
@@ -21,5 +21,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
         Task NotifyHealthEvent(ContainerHealthEventType healthEventType, Type source, string details);
 
         Task CreateBindMount(string sourcePath, string targetPath);
+
+        Task NotifyTriggersChanged(string contentHash);
     }
 }