+++ release_notes.md
@@ -5,3 +5,4 @@
 -->
 - Log enum names instead of integers in options formatting for `TimerTriggerPlatformOptions` (#11689)
 - Update WebHostWorkerRuntimeResolver to always read from IConfiguration variable value for FWR (#11720)
+- Notify mesh server of triggers update via `NotifyTriggersChanged` (#11690)